Moved rockets.js file to the rockets dir
