+++ src/redux/store.js
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
-import rocketsReducer from './rockets.js/rockets';
+import rocketsReducer from './rockets/rockets';
 import missionsReducer from './missions/missionsSlice';
 
 const store = configureStore({
